backend/requirements.txt
fastapi
uvicorn[standard]
anthropic==0.25.7
httpx==0.27.0
python-dotenv
requests
pydantic
